pyproject.toml
[tool.poetry]
name = "testcontainers_salt"
version = "0.1.1"
description = ""
authors = ["Nikolas Papaioannou <nikolas@hacky.software>"]
readme = "README.md"

[tool.poetyr.include]
include = ["resources"]

[tool.poetry.dependencies]
python = "^3.10"
testcontainers = "^4.10.0"
pyyaml = "^6.0.1"
pytest-testinfra = "^10.2.2"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
